+++ tests/IntegratedTest.php
@@ -15,9 +15,9 @@ class IntegratedTest extends TestCase
 
     public function test_get_existing_package()
     {
-        Artisan::call('packager:get', ['url' => 'https://github.com/Jeroen-G/laravel-packager', 'vendor' => 'MyVendor', 'name' => 'MyPackage']);
+        Artisan::call('packager:get', ['url' => 'https://github.com/Jeroen-G/packager-skeleton', 'vendor' => 'MyVendor', 'name' => 'MyPackage']);
 
-        $this->seeInConsoleOutput('Package created successfully!');
+        $this->seeInConsoleOutput('Package downloaded successfully!');
     }
 
     public function test_list_packages()
@@ -33,7 +33,7 @@ class IntegratedTest extends TestCase
         Artisan::call('packager:new', ['vendor' => 'MyVendor', 'name' => 'MyPackage']);
         $this->seeInConsoleOutput('MyVendor');
 
-        Artisan::call('packager:remove', ['vendor' => 'MyVendor', 'name' => 'MyPackage']);
+        Artisan::call('packager:remove', ['vendor' => 'MyVendor', 'name' => 'MyPackage', '--no-interaction' => true ]);
         $this->seeInConsoleOutput('Package removed successfully!');
     }
 }
