+++ tests/TestCase.php
@@ -2,11 +2,13 @@
 
 namespace JeroenG\Packager\Tests;
 
-use Illuminate\Contracts\Console\Kernel;
+use JeroenG\Packager\Tests\TestHelper;
 use Orchestra\Testbench\TestCase as TestBench;
 
 abstract class TestCase extends TestBench
 {
+    use TestHelper;
+
     /**
      * Setup before each test.
      *
@@ -25,22 +27,7 @@ abstract class TestCase extends TestBench
     {
         parent::tearDown();
 
-        $this->removeDir(getcwd().'/packages');
-    }
-
-    public function removeDir($path)
-    {
-        $files = array_diff(scandir($path), ['.', '..']);
-        foreach ($files as $file) {
-            if (is_dir("$path/$file")) {
-                $this->removeDir("$path/$file");
-            } else {
-                @chmod("$path/$file", 0777);
-                @unlink("$path/$file");
-            }
-        }
-
-        return rmdir($path);
+        $this->removeDir(base_path('packages'));
     }
 
     /**
@@ -51,17 +38,5 @@ abstract class TestCase extends TestBench
     protected function getPackageProviders($app)
     {
         return ['JeroenG\Packager\PackagerServiceProvider'];
-    }
-
-    protected function seeInConsoleOutput($expectedText)
-    {
-        $consoleOutput = $this->app[Kernel::class]->output();
-        $this->assertContains($expectedText, $consoleOutput, "Did not see `{$expectedText}` in console output: `$consoleOutput`");
-    }
-
-    protected function doNotSeeInConsoleOutput($unExpectedText)
-    {
-        $consoleOutput = $this->app[Kernel::class]->output();
-        $this->assertNotContains($unExpectedText, $consoleOutput, "Did not expect to see `{$unExpectedText}` in console output: `$consoleOutput`");
     }
 }
